Split up OMEMO into extra module
