+++ src/omemo/QXmppOmemoConfig.cmake.in
@@ -1,0 +1,12 @@
+# SPDX-FileCopyrightText: 2017 Niels Ole Salscheider <niels_ole@salscheider-online.de>
+#
+# SPDX-License-Identifier: CC0-1.0
+
+@PACKAGE_INIT@
+
+include(CMakeFindDependencyMacro)
+find_dependency(QXmpp)
+
+include("${CMAKE_CURRENT_LIST_DIR}/QXmppOmemo.cmake")
+
+check_required_components(QXmppOmemo)
