+++ src/omemo/QXmppOmemoDeviceBundle_p.h
