+++ src/omemo/QXmppOmemoDeviceElement_p.h
