+++ src/omemo/QXmppOmemoDeviceList_p.h
