+++ src/omemo/QXmppOmemoManager.cpp
