+++ src/omemo/QXmppOmemoMemoryStorage.cpp
