+++ tests/CMakeLists.txt
@@ -10,12 +10,17 @@ macro(add_simple_test TEST_NAME)
     add_executable(tst_${TEST_NAME} ${TEST_NAME}/tst_${TEST_NAME}.cpp ${ARGN})
     add_test(tst_${TEST_NAME} tst_${TEST_NAME})
     target_link_libraries(tst_${TEST_NAME} Qt${QT_VERSION_MAJOR}::Test qxmpp)
+    if(BUILD_OMEMO)
+        target_link_libraries(tst_${TEST_NAME} QXmppOmemo)
+    endif()
 endmacro()
 
 include_directories(${PROJECT_SOURCE_DIR}/src/base)
 include_directories(${PROJECT_SOURCE_DIR}/src/client)
+include_directories(${PROJECT_SOURCE_DIR}/src/omemo)
 include_directories(${PROJECT_SOURCE_DIR}/src/server)
 include_directories(${PROJECT_BINARY_DIR}/src/base)
+include_directories(${PROJECT_BINARY_DIR}/src/omemo)
 include_directories(${CMAKE_CURRENT_BINARY_DIR})
 
 add_simple_test(qxmpparchiveiq)
